Reverted some old changes so that original kirby is untouched, and GUI kirby exists entirely separate
