+++ robotkirby/bot.py
@@ -25,7 +25,7 @@ def make_client(bot: hikari.GatewayBot) -> tanjun.Client:
 
     client.load_modules('robotkirby.plugins.utilities')
     client.load_modules('robotkirby.plugins.log_message')
-    client.load_modules('robotkirby.plugins.bot_sentient')
+    client.load_modules('robotkirby.plugins.sentient')
     client.load_modules('robotkirby.plugins.optin')
     client.load_modules('robotkirby.plugins.delete')
     client.load_modules('robotkirby.plugins.wordcloud')
