+++ robotkirby/plugins/sentient.py
@@ -1,28 +1,47 @@
 import hikari
-from robotkirby.db.local_db_driver import Database
+import tanjun
+import typing
+from robotkirby.db.db_driver import Database
 import markovify
 
+component = tanjun.Component()
 
-def sentient(guild, member, channel, db: Database):
+
+@component.with_slash_command
+@tanjun.with_member_slash_option('member', 'user to imitate', default=None)
+@tanjun.with_channel_slash_option('channel', 'channel to imitate', default=None)
+@tanjun.as_slash_command('sentient', 'Imitate user')
+async def sentient(
+        ctx: tanjun.abc.Context,
+        member: typing.Optional[hikari.Member],
+        channel: typing.Optional[hikari.InteractionChannel],
+        db: Database = tanjun.inject(type=Database)
+) -> None:
+    if not db.check_read_permission(ctx.author):
+        await(ctx.respond('In order to use Robot Kirby, please opt in to data collection using the `/opt in` command. '
+                          'This will allow me to collect your messages so that I can build sentences from your data. '
+                          ':heart:'))
+        return
+
     match (member, channel):
         case (None, None):
-            prefix_str = f'**{guild["name"]}**'  # server name / dm title (maybe could use folder title or something)
-        case (member, None):
-            prefix_str = f'{member["name"]}'  # member name
-        case (None, channel):
-            prefix_str = f'{channel["name"]}'  # channel name
-        case (member, channel):
-            prefix_str = f'{member["name"]} in {channel["name"]}'  # member in channel
+            prefix_str = f'**{ctx.get_guild().name}**'
+        case (hikari.Member(), None):
+            prefix_str = f'{member.mention}'
+        case (None, hikari.InteractionChannel()):
+            prefix_str = f'{channel.mention}'
+        case (hikari.Member(), hikari.InteractionChannel()):
+            prefix_str = f'{member.mention} in {channel.mention}'
         case _:
-            raise Exception(f"Something is broken about this query.")
+            await ctx.respond(f"Something is broken about this query.")
+            return
 
-    print(f"Thinking about {prefix_str}...")
+    await ctx.respond(f"Thinking about {prefix_str}...")
 
     messages = db.get_messages(
         member=member,
-        guild=guild,
-        channel=channel,
-        text=None
+        guild=ctx.guild_id,
+        channel=channel
     )
 
     sentence = None
@@ -31,6 +50,11 @@ def sentient(guild, member, channel, db: Database):
         sentence = model.make_sentence()
 
     if sentence is not None:
-        return f"{prefix_str}:\n{sentence}"
+        await ctx.edit_initial_response(f"{prefix_str}:\n{sentence}")
     else:
-        return f"I don't have enough data for {prefix_str}."
+        await ctx.edit_initial_response(f"I don't have enough data for {prefix_str}.")
+
+
+@tanjun.as_loader
+def load(client: tanjun.abc.Client) -> None:
+    client.add_component(component.copy())
